Projects: set update_on_launch default to true (#17)
Follow the Principle of Least Astonishment and update the projects when the ref
doesn't contain a version.

examples:

- master => auto-update
- development => auto-update
- my-config-prod-1.2 => don't update

Default for cache-timeout to 30 so by the time someone makes a change and push it, then
run a job, there is a chance the change is picked up.

Even small cache allows should help in case of high load. Can be changed if
needed, after we load-test.